+++ R/addons-crossmap.R
@@ -1,0 +1,14 @@
+append_transpilers_for_crossmap <- function() {
+  append_transpilers("add-on", make_addon_transpilers(
+    "crossmap", "crossmap", make_options = function(options, fcn) {
+      keep <- intersect(names(formals(furrr::furrr_options)), names(options))
+      options <- options[keep]
+      options <- do.call(furrr::furrr_options, args = options)
+      options <- list(.options = options)
+      options
+    })
+  )
+
+  ## Return required packages
+  c("crossmap")
+}